Generates curves plots per peptide using the last cycle data

This function takes in the PW data (that includes all exposure times), list of peptides, and optional arguments of samples or groups names vector

krsa_curve_plot(data, peptides, byGroup = T, samples = NULL, groups = NULL)

Arguments

data

PW data (that includes all exposure times)

peptides

a vector of peptides to plot

byGroup

fit lines by group

samples

(optional) sample names

groups

(optional) group names

Value

ggplot object

See also

Examples

TRUE
#> [1] TRUE