Runs enrichr analysis using a list of PamChip peptide IDs or genes

This function takes in a list of PamChip peptide IDs or genes and geneset libraries names and connects to Enrichr api and returns enriched terms the geneset library name from Enrichr and userListId (id created by enrichr api) sand retruns tidy dataframe of all geneset terms in that library

Usage

krsa_enrichr(
  peptides = NULL,
  genes = NULL,
  libs = c("GO_Biological_Process_2021", "GO_Cellular_Component_2021",
    "GO_Molecular_Function_2021", "WikiPathway_2021_Human", "Reactome_2016",
    "KEGG_2021_Human", "BioPlanet_2019")
)

Arguments

peptides

a vector of PamChip peptide IDs

genes

(optional) a vector of HGNC symbols. This will override the list of peptides

libs

geneset libraries names from Enrichr. default: c("GO_Biological_Process_2021", "GO_Cellular_Component_2021","GO_Molecular_Function_2021", "WikiPathway_2021_Human","Reactome_2016", "KEGG_2021_Human", "BioPlanet_2019")

Value

df tidy dataframe of all geneset terms in that library

See also

Examples

TRUE
#> [1] TRUE